server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt==4.0.1",
    "chromium>=0.0.0",
    "dotenv>=0.9.9",
    "fastapi[standard]>=0.135.2",
    "openai-agents==0.6.2",
    "passlib>=1.7.4",
    "playwright>=1.58.0",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "python-jose[cryptography]>=3.5.0",
    "rich>=14.3.3",
    "slowapi>=0.1.9",
    "sqlmodel>=0.0.37",
]
